Ajout de MusicFileServlet Ajout de l'upload de fichiers audio Ajout de la suppression de fichiers audio Création de l'arborescence des fichiers et ajout de l'impact des manipulations des fichiers et dossiers sur l'arborescence

+++ src/java/bean/FolderBean.java
@@ -6,12 +6,15 @@
 package bean;
 
 import dao.FolderDAO;
+import dao.MusicFileDAO;
 import javax.inject.Named;
 import javax.enterprise.context.SessionScoped;
 import java.io.Serializable;
 import javax.faces.context.FacesContext;
 import model.Folder;
+import model.MusicFile;
 import model.User;
+import servlet.MusicFileServlet;
 
 /**
  *
@@ -24,14 +27,16 @@ public class FolderBean implements Serializable {
     private Folder currentFolder;
     private String currentFolderNewName;
     private String newFolderName;
-    
+
     private FolderDAO folderDAO;
+    private MusicFileDAO musicFileDAO;
 
     /**
      * Creates a new instance of FolderBean
      */
     public FolderBean() {
         folderDAO = new FolderDAO();
+        musicFileDAO = new MusicFileDAO();
     }
 
     public Folder getCurrentFolder() {
@@ -58,10 +63,18 @@ public class FolderBean implements Serializable {
         this.newFolderName = newFolderName;
     }
 
+    public String addMusicFile(MusicFile musicFile) {
+        currentFolder.musicFiles.add(musicFile);
+
+        return "folder?faces-redirect=true";
+    }
+
     public String open(Folder folder) {
         if (folder != null) {
             currentFolder = folder;
             currentFolderNewName = folder.name;
+
+            currentFolder.musicFiles = musicFileDAO.getMusicFilesByIdFolder(currentFolder.id);
 
             return "folder?faces-redirect=true";
         }
@@ -75,13 +88,16 @@ public class FolderBean implements Serializable {
     public String addFolder() {
         FacesContext context = FacesContext.getCurrentInstance();
         UserBean userBean = context.getApplication().evaluateExpressionGet(context, "#{userBean}", UserBean.class);
-        
+
         Folder folder = new Folder(userBean.getCurrentUser().id, newFolderName);
-        
+
         folder = folderDAO.createEntity(folder);
-        
+
+        MusicFileServlet.createDir(String.valueOf(userBean.getCurrentUser().id));
+        MusicFileServlet.createDir(userBean.getCurrentUser().id + "/" + folder.id);
+
         userBean.getCurrentUser().addFolder(folder);
-        
+
         newFolderName = null;
 
         return open(folder);
@@ -89,9 +105,9 @@ public class FolderBean implements Serializable {
 
     public String rename() {
         currentFolder.name = currentFolderNewName;
-        
+
         folderDAO.updateEntity(currentFolder);
-        
+
         return open(currentFolder);
     }
 
@@ -109,10 +125,12 @@ public class FolderBean implements Serializable {
         } else {
             folder = user.getFolderById(currentFolder.id - 1);
         }
-        
-        user.deleteFolderById(currentFolder.id);
-        
-        folderDAO.deleteEntity(currentFolder);
+
+        if (MusicFileServlet.deleteFolder()) {
+            user.deleteFolderById(currentFolder.id);
+
+            folderDAO.deleteEntity(currentFolder);
+        }
 
         return open(folder);
     }
